refactor: 解决了SonarLint检测的"Redundant casts should not be used"。
例如：this.igPositiveSentiment = (int) Math.round(((double) (fTotalPos + (float) iPosWords) + 0.45D) / (double) iPosWords);
完全使冗余的类型转换

+++ src/main/java/nju/SEIII/EASIEST/sentistrength/SentiStrength.java
@@ -578,19 +578,19 @@ public class SentiStrength {
                 System.out.print(
                         "Before training, binary accuracy: " + c.getClassificationTrinaryNumberCorrect() +
                                 " " +
-                                (float) c.getClassificationTrinaryNumberCorrect() / (float) corpusSize * 100.0F +
+                                c.getClassificationTrinaryNumberCorrect() / corpusSize * 100.0F +
                                 "%");
             } else {
                 System.out.print(
                         "Before training, trinary accuracy: " + c.getClassificationTrinaryNumberCorrect() +
                                 " " +
-                                (float) c.getClassificationTrinaryNumberCorrect() / (float) corpusSize * 100.0F +
+                                c.getClassificationTrinaryNumberCorrect() / corpusSize * 100.0F +
                                 "%");
             }
         } else if (c.options.bgScaleMode) {
             System.out.print(
                     "Before training, scale accuracy: " + c.getClassificationScaleNumberCorrect() + " " +
-                            (float) c.getClassificationScaleNumberCorrect() * 100.0F / (float) corpusSize +
+                            c.getClassificationScaleNumberCorrect() * 100.0F / corpusSize +
                             "% corr " + c.getClassificationScaleCorrelationWholeCorpus());
         } else {
             System.out.print(
